requirements.txt
Django==1.7.4
Unidecode==0.04.17
defusedxml==0.4.1
django-allauth==0.19.1
django-mptt==0.6.0
django-sendfile==0.3.6
djangorestframework==3.1.0
factory-boy==2.4.1
oauthlib==0.7.2
psycopg2==2.6
python3-openid==3.0.5
requests==2.5.1
requests-oauthlib==0.4.2
uWSGI==2.0.9
